+++ index.html
@@ -234,9 +234,8 @@
       </section>
     </template>
 
-    <!--Скрипты-->
-    <script src="./js/functions.js"></script>
-    <script src="./js/main.js"></script>
+    <!--Скрипт-->
+    <script src="./js/main.js" type="module"></script>
 
 </body>
 </html>
